backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.0.1
email-validator==2.2.0

SQLAlchemy==2.0.40
psycopg[binary]==3.2.9
alembic==1.15.2

passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0

slowapi==0.1.9

httpx==0.28.1

